+++ cmd/zetaclientd-supervisor/main.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"os/exec"
 	"os/signal"
+	"strings"
 	"syscall"
 	"time"
 
@@ -14,6 +15,7 @@ import (
 	"golang.org/x/sync/errgroup"
 
 	"github.com/zeta-chain/zetacore/app"
+	zetaos "github.com/zeta-chain/zetacore/pkg/os"
 	"github.com/zeta-chain/zetacore/zetaclient/config"
 )
 
@@ -37,7 +39,9 @@ func main() {
 	shutdownChan := make(chan os.Signal, 1)
 	signal.Notify(shutdownChan, syscall.SIGINT, syscall.SIGTERM)
 
-	hotkeyPassword, tssPassword, err := promptPasswords()
+	// prompt for all necessary passwords
+	titles := []string{"HotKey", "TSS", "Solana Relayer Key"}
+	passwords, err := zetaos.PromptPasswords(titles)
 	if err != nil {
 		logger.Error().Err(err).Msg("unable to get passwords")
 		os.Exit(1)
@@ -50,8 +54,6 @@ func main() {
 		os.Exit(1)
 	}
 	supervisor.Start(ctx)
-	// listen for SIGHUP to trigger a restart of zetaclientd
-	signal.Notify(supervisor.restartChan, syscall.SIGHUP)
 
 	shouldRestart := true
 	for shouldRestart {
@@ -66,7 +68,7 @@ func main() {
 		cmd.Stderr = os.Stderr
 		// must reset the passwordInputBuffer every iteration because reads are stateful (seek to end)
 		passwordInputBuffer := bytes.Buffer{}
-		passwordInputBuffer.Write([]byte(hotkeyPassword + "\n" + tssPassword + "\n"))
+		passwordInputBuffer.Write([]byte(strings.Join(passwords, "\n") + "\n"))
 		cmd.Stdin = &passwordInputBuffer
 
 		eg, ctx := errgroup.WithContext(ctx)
@@ -81,6 +83,7 @@ func main() {
 		})
 		eg.Go(func() error {
 			supervisor.WaitForReloadSignal(ctx)
+			cancel()
 			return nil
 		})
 		eg.Go(func() error {
@@ -88,8 +91,6 @@ func main() {
 				select {
 				case <-ctx.Done():
 					return nil
-				case sig := <-supervisor.restartChan:
-					logger.Info().Msgf("got signal %d, sending SIGINT to zetaclientd", sig)
 				case sig := <-shutdownChan:
 					logger.Info().Msgf("got signal %d, shutting down", sig)
 					shouldRestart = false