+++ e2e/e2etests/test_eth_deposit_refund.go
@@ -15,8 +15,8 @@ import (
 func TestEtherDepositAndCallRefund(r *runner.E2ERunner, args []string) {
 	require.Len(r, args, 1)
 
-	value, ok := big.NewInt(0).SetString(args[0], 10)
-	require.True(r, ok, "Invalid amount specified for TestEtherDepositAndCallRefund.")
+	// parse the deposit amount
+	value := parseBigInt(r, args[0])
 
 	evmClient := r.EVMClient