+++ e2e/e2etests/test_message_passing_zevm_to_evm_revert_fail.go
@@ -15,8 +15,8 @@ import (
 func TestMessagePassingZEVMtoEVMRevertFail(r *runner.E2ERunner, args []string) {
 	require.Len(r, args, 1)
 
-	amount, ok := big.NewInt(0).SetString(args[0], 10)
-	require.True(r, ok, "Invalid amount specified for TestMessagePassingZEVMtoEVMRevertFail.")
+	// parse the amount
+	amount := parseBigInt(r, args[0])
 
 	// Deploying a test contract not containing a logic for reverting the cctx
 	testDappNoRevertAddr, tx, testDappNoRevert, err := testdappnorevert.DeployTestDAppNoRevert(