+++ x/crosschain/keeper/gas_payment.go
@@ -169,11 +169,19 @@ func (k Keeper) PayGasInERC20AndUpdateCctx(
 	if _, found := k.zetaObserverKeeper.GetSupportedChainFromChainID(ctx, chainID); !found {
 		return observertypes.ErrSupportedChains
 	}
+
 	// get gas params
 	gas, err := k.ChainGasParams(ctx, chainID)
 	if err != nil {
 		return cosmoserrors.Wrap(types.ErrCannotFindGasParams, err.Error())
 	}
+
+	// with V2 protocol, reverts on connected chains can eventually call a onRevert function which can require a higher gas limit
+	if cctx.ProtocolContractVersion == types.ProtocolContractVersion_V2 && cctx.RevertOptions.CallOnRevert &&
+		!cctx.RevertOptions.RevertGasLimit.IsZero() {
+		gas.GasLimit = cctx.RevertOptions.RevertGasLimit
+	}
+
 	outTxGasFee := gas.GasLimit.Mul(gas.GasPrice).Add(gas.ProtocolFlatFee)
 	// get address of the zrc20
 	fc, found := k.fungibleKeeper.GetForeignCoinFromAsset(ctx, cctx.InboundParams.Asset, chainID)